Faster docs (#3189)
* Enables gzip when serving the docs site

This should result in a 50-60% improvement in payload size for users
browsing the docs site.
* Precompile docs search index

Uses lunrjs (updated from 0.7.2 to 2.3.7) to pre-build the search index
for the docs site. Pre-building this gives us two big wins:
1) The ~2mB raw / ~500kB gzipped hugo index no longer needs to be transferred
over the wire.
2) The browser doesn't need to build the big search-index on every page load,
removing the need for  about ~2-3s of blocking js, which makes the site feel
a lot more performant.
* Move changelog to latest
* Update to latest version of solo theme

Enables new lunr search
* Use 'current' for default image
* Update changelog to resolve issue

This PR now solves everything in the linked issue