gateway2/routes: prioritize regex over prefix and ignore regex lengths (#9473)
This change updates the route sorting such that regex based matchers
are prioritized over prefix based matchers to be consistent with
Gloo gateway v1. The Gateway API states that the precedence of
regex matches is implementation specific, so this change conforms
to the API requirements.

Further, it avoids prioritizing regex based matchers based on their
lengths as this is rather abitrary. E.g., /a/b.*/.* is longer than
/a/b./c but less specific, so it's relative order should be lower
in the sorted order but if prioritized based on length it would
appear before a more specific match. If there is a tie among
regex routes, the other precedence rules govern the relative
ordering among them.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>
Co-authored-by: soloio-bulldozer[bot] <48420018+soloio-bulldozer[bot]@users.noreply.github.com>