gateway2/route-options: merge ExtensionRef override with targetRef attachment (#9486)
Implements merging of RouteOptions attached via ExtensionRef filter
and targetRef based policy attachment. ExtensionRef based options
have higher priority than targetRef based options, such that if
the ExtensionRef based option sets a field, it cannot be overridden
by a policy using targetRef. However, unset fields on a policy
selected by ExtensionRef may be overridden by targetRef based options.

The merging semantics is already followed when policies are merged
along a delegation chain, so this change is required for consistency.

The core of this change to merge policies in order of their priority
within the query resolver itself because the resolver is responsible
for recursively resolving the policies for delegated routes. Hence,
existing lookup of ExtensionRef based options are moved into the
query API that also handles the merge.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>