gateway2/delegation: enable optional parent matcher inheritance (#9513)
In the current design, a child route must have route matchers
that match the parent's to be considered in the delegation chain.

There's a request to enable matcher inheritance from the parent
such that instead of child routes needing to specify paths
that include the parent's path as prefix, they instead inherit
the parent prefix, headers, query params, etc. The result
is that the paths specified by child routes are relative
to the parent paths. Headers and query params are merged
from the parent into the child while giving preference to
parent matchers in case of conflicts.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>